Add safety checks when loading from m3u or cmd
This PR adds checks for empty string or path. check if file exists and
if file has a supported extension when loading for CMD or M3U format.
This prevents users from adding invalid files in m3u or cmd format and
accidentally load them in core causing them to get corrupted.